feat(demo): one-click sign-in with the shared demo accounts (#966)
On the demo instance the sign-in page now shows the three shared
accounts as buttons (Admin/Mentor/Mentee) that sign in directly, so a
visitor no longer copies credentials from /demo by hand.

src/app/auth/signin/page.tsx becomes a thin server wrapper: it resolves
the server-only IS_DEMO_MODE flag and passes DEMO_ACCOUNTS +
DEMO_PASSWORD to the client form (now SignInClient.tsx) as a prop. On
every non-demo instance the prop is null and the page renders exactly
as before — guarded by a new e2e test asserting the panel is absent.
The Safari session-settle poll moved into settleAndRedirect(), shared
by the form submit and the quick-login path.

Verified locally: with DEMO_MODE=true the SSR HTML carries the panel
and all three role buttons; without it, e2e shows no panel and the
existing login-security spec still passes.
