+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "internship-crm",
-  "version": "0.75.0-beta",
+  "version": "0.76.0-beta",
   "private": true,
   "license": "AGPL-3.0-or-later",
   "author": "Mehmet Erşahin",
