+++ src/i18n/dictionaries.ts
@@ -1166,6 +1166,8 @@ const en = {
     title: 'Live interactive demo',
     subtitle: 'Sign in with any of the accounts below and use the app for real — move candidates through the pipeline, log interactions, create projects. Nothing here is anyone’s actual data.',
     credentialsTitle: 'Sign in as',
+    quickTitle: 'Demo accounts — one-click sign-in',
+    quickHint: 'Shared accounts on synthetic data; the demo resets twice a day.',
     colRole: 'Role',
     colEmail: 'Email',
     colPassword: 'Password',
@@ -3782,6 +3784,8 @@ const tr: Dict = {
     title: 'Canlı etkileşimli demo',
     subtitle: 'Aşağıdaki hesaplardan biriyle giriş yapın ve uygulamayı gerçekten kullanın — adayları süreçte ilerletin, görüşme kaydı girin, proje oluşturun. Buradaki hiçbir şey gerçek bir kişinin verisi değil.',
     credentialsTitle: 'Şununla giriş yap',
+    quickTitle: 'Demo hesapları — tek tıkla giriş',
+    quickHint: 'Sentetik veri üzerinde ortak hesaplar; demo günde iki kez sıfırlanır.',
     colRole: 'Rol',
     colEmail: 'E-posta',
     colPassword: 'Şifre',
@@ -6394,6 +6398,8 @@ const de: Dict = {
     title: 'Interaktive Live-Demo',
     subtitle: 'Melden Sie sich mit einem der folgenden Konten an und nutzen Sie die App wirklich — Kandidaten durch die Pipeline bewegen, Gespräche erfassen, Projekte anlegen. Nichts hier sind echte Daten einer Person.',
     credentialsTitle: 'Anmelden als',
+    quickTitle: 'Demo-Konten — Anmeldung mit einem Klick',
+    quickHint: 'Gemeinsame Konten auf synthetischen Daten; die Demo wird zweimal täglich zurückgesetzt.',
     colRole: 'Rolle',
     colEmail: 'E-Mail',
     colPassword: 'Passwort',
